Feat: Add ArchiveListFiltering

+++ app/src/main/java/com/likefirst/btos/ui/archive/ArchiveListFragment.kt
@@ -24,33 +24,56 @@ class ArchiveListFragment : BaseFragment<FragmentArchiveListBinding>(FragmentArc
     companion object{
         var requiredPageNum = 1
         var lastDate = ""
+        var datePickerFlag = true
     }
 
     override fun initAfterBinding() {
         //companion object 초기화
         requiredPageNum = 1
         lastDate = ""
-        
-        //datePicker 기능 설정
-        binding.archiveListToolbar.archiveListPeriodIv.setOnClickListener {
-            it.isSelected = !it.isSelected
-            val periodDialog = ArchiveListPeriodDialog()
-            periodDialog.setStyle(BottomSheetDialogFragment.STYLE_NORMAL, R.style.ArchiveDatePickerStyle)
-            periodDialog.show(childFragmentManager, periodDialog.tag)
-        }
-        initDiaryList()
+
+        val mAdapter = ArchiveListRVAdapter(requireContext())
+        initDatePicker(mAdapter)
+        initDiaryList(mAdapter)
     }
 
-    fun initDiaryList(){
-        val mAdapter = ArchiveListRVAdapter(requireContext())
-        val mDecoration = ArchiveListItemDecoration()
-        mDecoration.setSize(requireContext())
+    fun initDatePicker(mAdapter: ArchiveListRVAdapter){
+        //datePicker 기능 설정
+        binding.archiveListToolbar.archiveListPeriodIv.setOnClickListener {
+            if(it.isSelected){
+                val query = HashMap<String, String>()
+                reLoadDiaryList(mAdapter, query)
+                it.isSelected = !it.isSelected
+            } else {
+                if(datePickerFlag){
+                    val datePickerDialog = ArchiveListPeriodDialog.newInstance()
+                    datePickerDialog.setDatePickerClickListener(object : ArchiveListPeriodDialog.DatePickerClickListener{
+                        override fun onDatePicked(dateFrom: String, dateTo: String) {
+                            val query = HashMap<String, String>()
+                            query["startDate"] = dateFrom
+                            query["endDate"] = dateTo
+//                            binding.archiveListRv.clearOnScrollListeners()
+//                            mAdapter.clearList()
+                            reLoadDiaryList(mAdapter, query)
+                            it.isSelected = !it.isSelected
+                        }
+                    })
+                    datePickerDialog.setStyle(BottomSheetDialogFragment.STYLE_NORMAL, R.style.ArchiveDatePickerStyle)
+                    datePickerDialog.show(childFragmentManager, datePickerDialog.tag)
+                    datePickerFlag = false
+                }
+            }
+        }
+    }
+
+    fun reLoadDiaryList(mAdapter: ArchiveListRVAdapter, query: HashMap<String, String>){
+        binding.archiveListRv.clearOnScrollListeners()
         binding.archiveListRv.addOnScrollListener(object : RecyclerView.OnScrollListener(){
             override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
                 super.onScrolled(recyclerView, dx, dy)
 
                 // 스크롤이 끝에 도달했는지 확인
-                if (!binding.archiveListRv.canScrollVertically(1)) {
+                if (!binding.archiveListRv.canScrollVertically(1) && !mAdapter.isDiaryEmpty()) {
                     if(requiredPageNum == 0){
                         // 더이상 불러올 페이지가 없으면 스크롤 리스너 clear
                         mAdapter.deleteLoading()
@@ -58,7 +81,36 @@ class ArchiveListFragment : BaseFragment<FragmentArchiveListBinding>(FragmentArc
                         binding.archiveListRv.clearOnScrollListeners()
                     } else {
                         // 다음 페이지 불러오기
-                        loadDiaryList(requiredPageNum, mAdapter)
+                        loadDiaryList(requiredPageNum, mAdapter, query)
+                    }
+                }
+            }
+        })
+        requiredPageNum = 1
+        lastDate = ""
+        mAdapter.clearList()
+        Log.d("requiredPageNum", requiredPageNum.toString())
+        loadDiaryList(requiredPageNum, mAdapter, query)
+    }
+
+    fun initDiaryList(mAdapter : ArchiveListRVAdapter){
+        val query = HashMap<String, String>()
+        val mDecoration = ArchiveListItemDecoration()
+        mDecoration.setSize(requireContext())
+        binding.archiveListRv.addOnScrollListener(object : RecyclerView.OnScrollListener(){
+            override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
+                super.onScrolled(recyclerView, dx, dy)
+
+                // 스크롤이 끝에 도달했는지 확인
+                if (!binding.archiveListRv.canScrollVertically(1) && !mAdapter.isDiaryEmpty()) {
+                    if(requiredPageNum == 0){
+                        // 더이상 불러올 페이지가 없으면 스크롤 리스너 clear
+                        mAdapter.deleteLoading()
+                        mAdapter.notifyItemRemoved(mAdapter.itemCount)
+                        binding.archiveListRv.clearOnScrollListeners()
+                    } else {
+                        // 다음 페이지 불러오기
+                        loadDiaryList(requiredPageNum, mAdapter, query)
                     }
                 }
             }
@@ -67,15 +119,16 @@ class ArchiveListFragment : BaseFragment<FragmentArchiveListBinding>(FragmentArc
             adapter = mAdapter
             overScrollMode = RecyclerView.OVER_SCROLL_NEVER
             layoutManager = LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
-            addItemDecoration(mDecoration)
+            if (itemDecorationCount == 0){
+                addItemDecoration(mDecoration)
+            }
         }
-        loadDiaryList(requiredPageNum, mAdapter)
+        loadDiaryList(requiredPageNum, mAdapter, query)
     }
 
-    fun loadDiaryList(pageNum : Int, adapter : ArchiveListRVAdapter){
-        Log.d("loadDiaryList", "loadDiaryList")
+    fun loadDiaryList(pageNum : Int, adapter : ArchiveListRVAdapter, query : HashMap<String,String>){
         // 검색조건에 따라 쿼리스트링 생성해서 전달
-        val query = HashMap<String,String>()
+        Log.d("requiredPageNum", requiredPageNum.toString())
         val archiveListService = ArchiveListService()
         archiveListService.setArchiveListView(this)
         archiveListService.getList(getUserIdx(), pageNum, query, adapter)
@@ -89,6 +142,7 @@ class ArchiveListFragment : BaseFragment<FragmentArchiveListBinding>(FragmentArc
         pageInfo: ArchiveListPageInfo,
         adapter: ArchiveListRVAdapter
     ) {
+        Log.d("archiveSuccess", "success!!!!!!")
         if(pageInfo.currentPage != 1){
             adapter.deleteLoading()
         }
@@ -104,7 +158,9 @@ class ArchiveListFragment : BaseFragment<FragmentArchiveListBinding>(FragmentArc
                 diaryList.add(item.diaryList[i])
             }
         }
-        diaryList.add(0)
+        if (pageInfo.hasNext){
+            diaryList.add(0)
+        }
         adapter.apply{
             addDiaryList(diaryList)
             notifyItemRangeInserted((pageInfo.currentPage-1)*(result.size), result.size + 1)
@@ -114,9 +170,10 @@ class ArchiveListFragment : BaseFragment<FragmentArchiveListBinding>(FragmentArc
         } else {
             requiredPageNum = 0
         }
+        Log.d("requiredPageNum++", requiredPageNum.toString())
     }
 
     override fun onArchiveListFailure(code: Int) {
-        
+
     }
 }
